requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asyncio==4.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
distro==1.9.0
dotenv==0.9.9
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
numpy==2.4.2
polling2==0.5.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sounddevice==0.5.5
speechmatics-rt==0.5.3
speechmatics-voice==0.2.8
tenacity==8.2.3
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
